+++ docs/changelog.rst
@@ -1,6 +1,10 @@
 =========
 Changelog
 =========
+* :release:`9.0.0 <2022-12-20>`
+* :feature:`372` Acoustic transfer propagates well properties
+* :feature:`371` Added new container type: `384-corning-4512-round-lv`
+
 * :release:`8.0.1 <2022-12-07>`
 * :feature:`365` Update autopick method signature to allow for multiple pick groups per instruction